cleanup: move stuff around a bit
