+++ distribution/lib/Standard/Table/0.0.0-dev/src/Data/Table.enso
@@ -67,9 +67,9 @@ polyglot java import org.enso.base.ObjectComparator
 polyglot java import org.enso.table.data.index.MultiValueIndex
 polyglot java import org.enso.table.data.mask.OrderMask
 polyglot java import org.enso.table.data.table.Column as Java_Column
-polyglot java import org.enso.table.data.table.join.Between as Java_Join_Between
-polyglot java import org.enso.table.data.table.join.Equals as Java_Join_Equals
-polyglot java import org.enso.table.data.table.join.EqualsIgnoreCase as Java_Join_Equals_Ignore_Case
+polyglot java import org.enso.table.data.table.join.conditions.Between as Java_Join_Between
+polyglot java import org.enso.table.data.table.join.conditions.Equals as Java_Join_Equals
+polyglot java import org.enso.table.data.table.join.conditions.EqualsIgnoreCase as Java_Join_Equals_Ignore_Case
 polyglot java import org.enso.table.data.table.join.LookupJoin
 polyglot java import org.enso.table.data.table.Table as Java_Table
 polyglot java import org.enso.table.error.TooManyColumnsException
