+++ std-bits/table/src/main/java/org/enso/table/data/table/join/LookupJoin.java
@@ -9,6 +9,7 @@ import org.enso.table.data.index.UnorderedMultiValueKey;
 import org.enso.table.data.mask.OrderMask;
 import org.enso.table.data.table.Column;
 import org.enso.table.data.table.Table;
+import org.enso.table.data.table.join.conditions.Equals;
 import org.enso.table.error.NonUniqueLookupKey;
 import org.enso.table.error.NullValuesInKeyColumns;
 import org.enso.table.error.UnmatchedRow;
@@ -17,7 +18,6 @@ import org.enso.table.util.ConstantList;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.IntStream;
 
 public class LookupJoin {
