+++ std-bits/table/src/main/java/org/enso/table/data/table/join/conditions/HashableCondition.java
@@ -1,4 +1,4 @@
-package org.enso.table.data.table.join;
+package org.enso.table.data.table.join.conditions;
 
 public sealed interface HashableCondition extends JoinCondition permits Equals, EqualsIgnoreCase {
 }
