+++ std-bits/table/src/main/java/org/enso/table/data/table/join/conditions/JoinCondition.java
@@ -1,3 +1,3 @@
-package org.enso.table.data.table.join;
+package org.enso.table.data.table.join.conditions;
 
 public sealed interface JoinCondition permits HashableCondition, Between {}
